split on 2 different components
